creating offer endpoints: Create and Get All

+++ src/main/java/com/vivomoney/dtos/OfferDTO.java
@@ -1,6 +1,4 @@
 package com.vivomoney.dtos;
 
-import java.math.BigDecimal;
-
-public record OfferDTO(BigDecimal taxa_mensal, Long valor, Long numero_parcelas) {
+public record OfferDTO(double taxa_mensal, Long valor, Long numero_parcelas) {
 }
